requirements.txt
asgiref==3.5.2
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
dj-rest-auth==2.2.5
Django==3.2.14
django-allauth==0.51.0
django-braces==1.15.0
django-cors-headers==3.13.0
django-import-export==2.8.0
django-oauth-toolkit==2.1.0
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
firebase-admin==5.2.0
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.0
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.48.0
grpcio-status==1.48.0
httplib2==0.20.4
idna==3.3
jwcrypto==1.3.1
MarkupPy==1.14
msgpack==1.0.4
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
proto-plus==1.20.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tablib==3.2.1
uritemplate==4.1.1
urllib3==1.26.10
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0